[htp] Import HVX exp/inverse and dma-queue primitives from ggml-hexagon
Import 11 files from llama.cpp ggml-hexagon (MIT), split by hardware
unit: scalar-core utilities hex-{common,dump,fastdiv,utils}.h under
htp/hex/, the user-DMA queue dma-queue.{c,h} under htp/dma/, and the
HVX vector primitives hvx-{base,exp,floor,inverse,types}.h under
htp/hvx/. Beyond the 7 planned files, hvx-types.h, hex-fastdiv.h,
hex-dump.h and hex-common.h come along as hard include dependencies
with no ggml coupling. Each file keeps the upstream code and gains an
SPDX MIT tag with the import origin and copyright inside a single
doxygen header block. Add the corresponding NOTICE entry and the
hex/dma include paths and dma source glob to build_sim_test.sh.

Local modifications to the imported code, noted in the file headers:
- dma-queue.h: hex-profile.h include dropped (not imported); the
  profiling hooks are stubbed as no-ops with matching signatures.
- hex-common.h: hmx_ceil_div renamed to hex_ceil_div so identifiers
  in the file follow their actual unit prefix (the helper is generic
  scalar code, not HMX-specific).
- hex-dump.h relies on <stdio.h> arriving transitively via HAP_farf.h;
  left as upstream, noted here in case a future SDK drops that pull-in.

Add test_exp to the sim harness: 4096 fp32 inputs in [-20, 4) run
through the imported hvx_exp_f32 against expf within rtol 1e-3 /
atol 1e-6. This is the first cmp_f caller, which exposed a latent
-Wformat error in sim_test_util.h (uint32_t is unsigned long under
hexagon-clang); fixed with an (unsigned) cast.

Verified on hexagon-sim v75: SIM_TEST exp PASS, exit 0.

+++ test/hexagon/sim/sim_test_util.h
@@ -31,7 +31,7 @@ static inline int cmp_f(const char *tag, const float *ref, const float *got,
     }
   }
   if (worst > 0.f) {
-    printf("SIM_TEST %s FAIL i=%u ref=%f got=%f\n", tag, wi, ref[wi], got[wi]);
+    printf("SIM_TEST %s FAIL i=%u ref=%f got=%f\n", tag, (unsigned)wi, ref[wi], got[wi]);
     return 1;
   }
   return 0;
